+++ conftest.py
@@ -1,0 +1,1 @@
+pytest_plugins = ["ghost_pii.pytest_plugin"]